Restructured server manger's scripts
