+++ server_manager/src/lib/globals.js
